Refactor: Separate prediction services into dedicated files
The `services/prediction_service.py` file, which originally contained
functionalities for both compatibility scoring and user-based recommendations,
has been refactored.

The functionalities are now split into two distinct files:
- `services/compatibility_service.py`: Contains the `get_compatibility_scores`
  function, responsible for predicting compatibility between you and
  programs/activities.
- `services/recommendation_service.py`: Contains
  `generate_user_based_recommendations`, `get_recommendations_for_user`,
  and `get_top_recommended_activities` functions, which handle
  collaborative filtering based recommendations.

Import statements in controller files (`dashboard_routes.py` and
`program_controller.py`) have been updated to reflect these new locations.
The original `services/prediction_service.py` file has been removed.

This change improves modularity and separation of concerns within the
services layer without altering the existing functionality.
